# Improve performance of StringBuilder